bug 1328989: Move docs/macros to dashboards/macros
