+++ kuma/dashboards/jinja2/dashboards/macros.html
@@ -1,4 +1,4 @@
-{% extends "wiki/base.html" %}
+{% extends "base.html" %}
 {% set styles = ("dashboards",) %}
 {% set title = _("Active macros") %}
 {% set github_url = "https://github.com/mozilla/kumascript/tree/master/macros/" %}
