+++ kuma/wiki/urls.py
@@ -108,9 +108,6 @@ urlpatterns = [
     url(r'^/templates$',
         views.list.templates,
         name='wiki.list_templates'),
-    url(r'^/macros$',
-        views.list.macros,
-        name='wiki.list_macros'),
     url(r'^/tags$',
         views.list.tags,
         name='wiki.list_tags'),
